+++ src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
 
@@ -23,6 +23,8 @@ import { AppComponent } from './app.component';
   ],
   imports: [
     BrowserModule,
+    FormsModule,
+    ReactiveFormsModule,
     ProductsModule,
     CartModule,
     SharedModule,
@@ -31,7 +33,8 @@ import { AppComponent } from './app.component';
     CoreModule,
     HttpClientModule,
 
-    AppRoutingModule
+    AppRoutingModule,
+
   ],
   bootstrap: [AppComponent],
   providers: [httpInterceptorProviders]