+++ src/app/layout/components/order/models/order.model.ts
@@ -2,11 +2,13 @@ import { CartProduct } from './../../../../cart/models/cart-product.model';
 
 export class OrderModel {
     constructor(
-        public firstName: string,
-        public lastName: string,
-        public productList: CartProduct[],
-        public id?: number
-    ) {
-        this.id = id || +Date.now();
-    }
+        public firstName: string = '',
+        public lastName: string = '',
+        public email = '',
+        public phone = null,
+        public pickup = false,
+        public address = [''],
+        public productList: CartProduct[] = [],
+        public id = +Date.now()
+    ) { }
 }