+++ src/app/layout/layout.module.ts
@@ -1,15 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { ValidatorsModule } from './../validators/validators.module';
+import { ProcessOrderServiceModule } from './../process-order/process-order-service.module';
+
 
 import { AboutComponent, PathNotFoundComponent,
     OrderComponent, LoginComponent } from './components';
 
+import { ProcessOrderComponent } from './../process-order/component/process-order/process-order.component';
+
+
 @NgModule({
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, ReactiveFormsModule, ValidatorsModule, ProcessOrderServiceModule],
   providers: [],
   declarations: [AboutComponent, PathNotFoundComponent, OrderComponent,
-  LoginComponent],
+  LoginComponent, ProcessOrderComponent],
   exports: []
 })